refactor: share hands-free voice review loop via the kernel
Moves HandsFreeReviewController and the spoken-rating parsing out of the
Android companion into src/kernel/recall/voice-review.ts so the desktop app
and the iOS companion can drive the identical loop. mobile/src/voice.ts stays
as the mobile entry point and re-exports it.

Adds the engine-tiering logic voice mode needs on every surface: a user
preference (device-only / device-first / quality-first), per-capability
resolution against what the device and the model registry can actually do, and
an explicit reason so a surface can tell the learner when speech left the
device.
